Resolve quoting problems with sending code to urb
Sometimes when sending code to urb, bash would attempt to interpret
the code instead of just sending it along. I tried multiple different
techniques for fixing this and had to fall back to the emacs
shell-quote-argument method. The downside is that the exact command
send to urb is difficult to examine, but it does maintain
asynchronicity which is important for me not throwing my computer out
the window. The difficulty of examining the command should not be a
problem in day to day use as you will be examining your code rather
than the input to urb.